+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 
 ## Other  ?? Technical stuff, what happened behind the scene
 -->
+## Fixes
+- Fixed some preference not being saved properly
+
 ## Other
 - Update dependency co.touchlab:kermit to v2.0.4
 - Update lifecycle to v2.8.4